🛠결과_일단 단답형,주관식 진행 및 모듈화,,
